Popup: Replacing version in react-internal with version from react-next.
