+++ packages/react-internal/src/components/Panel/Panel.types.ts
@@ -20,7 +20,7 @@ export interface IPanel {
   /**
    * Forces the panel to dismiss.
    */
-  dismiss: (ev?: React.KeyboardEvent<HTMLElement>) => void;
+  dismiss: (ev?: React.KeyboardEvent<HTMLElement> | KeyboardEvent) => void;
 }
 
 /**
@@ -98,7 +98,7 @@ export interface IPanelProps extends React.HTMLAttributes<PanelBase> {
    * A callback function for when the panel is closed, before the animation completes.
    * If the panel should NOT be dismissed based on some keyboard event, then simply call ev.preventDefault() on it
    */
-  onDismiss?: (ev?: React.SyntheticEvent<HTMLElement>) => void;
+  onDismiss?: (ev?: React.SyntheticEvent<HTMLElement> | KeyboardEvent) => void;
 
   /**
    * A callback function which is called **after** the Panel is dismissed and the animation is complete.
