--- packages/react-next/src/components/Popup/index.ts
@@ -1,2 +1,0 @@
-export * from './Popup';
-export * from './Popup.types';
